fix bugs, implements environmental data and community reports.

+++ src/main/java/com/eco/track/eco_tracking_system/service/UserProfileService.java
@@ -9,6 +9,7 @@ import org.springframework.validation.BindingResult;
 import com.eco.track.eco_tracking_system.config.security.JwtService;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 import com.eco.track.eco_tracking_system.entity.UserProfile;
@@ -21,15 +22,15 @@ import com.eco.track.eco_tracking_system.repository.ProfileFollowersRepository;
 
 import com.eco.track.eco_tracking_system.dto.UserProfileDTO;
 
-import com.eco.track.eco_tracking_system.request.UserProfileRequest;
-import com.eco.track.eco_tracking_system.response.UserProfileResponse;
-import com.eco.track.eco_tracking_system.response.UserProfilesResponse;
+import com.eco.track.eco_tracking_system.response.RecordResponse;
 import com.eco.track.eco_tracking_system.response.GenericResponse;
+import com.eco.track.eco_tracking_system.request.UserProfileRequest;
 
 import com.eco.track.eco_tracking_system.util.Helper;
 
 import com.eco.track.eco_tracking_system.exception.ExceptionType.NotFoundException;
 import com.eco.track.eco_tracking_system.exception.ExceptionType.ValidationException;
+import com.eco.track.eco_tracking_system.exception.ExceptionType.UnauthorizedException;
 
 @Service
 @RequiredArgsConstructor
@@ -83,18 +84,33 @@ public class UserProfileService
     public GenericResponse updateProfile(
         long id,
         UserProfileRequest request,
-        BindingResult result
+        BindingResult result,
+        String token
     ) throws
         ValidationException,
-        NotFoundException
+        NotFoundException,
+        UnauthorizedException
     {
         Helper.fieldsValidate(result);
+
+        String username = jwtService.extractUsername(token);
+
+        var user = userRepository.findByUsername(username)
+            .orElseThrow(() -> new NotFoundException("user not found"));
 
         var userProfile = userProfileRepository.findById(id)
             .orElseThrow(() -> new NotFoundException("profile not found"));
 
         var topic = topicRepository.findById(request.getParsedTopicID())
             .orElseThrow(() -> new NotFoundException("topic not found"));
+
+        int size = user.getUserProfiles()
+            .stream()
+            .filter(profile -> Objects.equals(profile.getProfileID(), userProfile.getProfileID()))
+            .toList().size();
+
+        if(size == 0)
+            throw new UnauthorizedException("unauthorized user");
 
         userProfile.setTopic(topic);
         userProfile.setFullName(request.getFullName());
@@ -108,7 +124,28 @@ public class UserProfileService
             .build();
     }
 
-    public UserProfilesResponse getProfiles(
+    public RecordResponse<List<UserProfileDTO>> getAllProfiles()
+    {
+        List<UserProfile> userProfiles = userProfileRepository.findAll();
+
+        List<UserProfileDTO> userProfileDTOS = userProfiles
+            .stream()
+            .map(profile -> UserProfileDTO
+                    .builder()
+                    .id(profile.getProfileID())
+                    .topicID(profile.getTopic().getTopicID())
+                    .fullName(profile.getFullName())
+                    .build()
+            )
+            .toList();
+
+        return RecordResponse
+            .<List<UserProfileDTO>>builder()
+            .data(userProfileDTOS)
+            .build();
+    }
+
+    public RecordResponse<List<UserProfileDTO>> getProfiles(
         String token
     ) throws NotFoundException
     {
@@ -130,13 +167,13 @@ public class UserProfileService
             )
             .toList();
 
-        return UserProfilesResponse
-            .builder()
-            .userProfiles(userProfileDTOS)
+        return RecordResponse
+            .<List<UserProfileDTO>>builder()
+            .data(userProfileDTOS)
             .build();
     }
 
-    public UserProfileResponse getProfile(
+    public RecordResponse<UserProfileDTO> getProfile(
         long id
     ) throws NotFoundException
     {
@@ -150,19 +187,35 @@ public class UserProfileService
             .fullName(profile.getFullName())
             .build();
 
-        return UserProfileResponse
-            .builder()
-            .userProfile(profileDTO)
+        return RecordResponse
+            .<UserProfileDTO>builder()
+            .data(profileDTO)
             .build();
     }
 
     @Transactional
     public GenericResponse deleteProfile(
-        long id
-    ) throws NotFoundException
+        long id,
+        String token
+    ) throws
+        NotFoundException,
+        UnauthorizedException
     {
+        String username = jwtService.extractUsername(token);
+
+        var user = userRepository.findByUsername(username)
+            .orElseThrow(() -> new NotFoundException("user not found"));
+
         var profile = userProfileRepository.findById(id)
             .orElseThrow(() -> new NotFoundException("profile not found"));
+
+        int size = user.getUserProfiles()
+            .stream()
+            .filter(userProfile -> Objects.equals(profile.getProfileID(), userProfile.getProfileID()))
+            .toList().size();
+
+        if(size == 0)
+            throw new UnauthorizedException("unauthorized user");
 
         userProfileRepository.delete(profile);
